refactor : CLI structure and add command options handling
